Refractored FixBookUi
